Fixed from int id to long id in PersonController

+++ src/main/java/kea/group4/backend/api/PersonController.java
@@ -36,13 +36,13 @@ public class PersonController {
 
     //Update
     @PutMapping("/{id}")
-    public PersonResponse editPerson(@RequestBody PersonRequest body, @PathVariable int id){
+    public PersonResponse editPerson(@RequestBody PersonRequest body, @PathVariable long id){
         return personService.editPerson(body, id);
     }
 
     //Delete
     @DeleteMapping("/{id}")
-    public void deletePerson(@PathVariable int id){
+    public void deletePerson(@PathVariable long id){
         personService.deletePerson(id);
     }
 }
